[FragmentLoader] Only use max video tags available to fix a Fragment
We assumed that current and emergency Fragments would have headers at
same PTS, but it seems this may not be the case. Fortunately if we are
fixing a Fragment all we need are video tags up to that PTS, regardless
of others.

We add some extra checking to be sure nothing strange happens.